requirements.txt
aiogram==3.4.1
pillow==10.2.0
python-dotenv==1.0.1
